Add more complex example of good / bad conditionals
IMHO a condition like `((score > 25) && (score < 50))` is simple enough
that it can stay as-is (assuming the intent is clear from the
surrounding code and / or comments).

However I've seen (and written) more complex expressions like the
updated example which in my opinion benefit a lot more from being
separated out as distinct conditions.